Continue cleanup and move all netoworking to App class

+++ src/main/java/uk/ac/ed/inf/powergrab/ChargingStation.java
@@ -22,8 +22,8 @@ public class ChargingStation {
 	}
 	
 	public void charge(Drone drone) {
-		double newCoins = drone.getCoins() + coins;
-		double newPower = drone.getPower() + power;
+		float newCoins = (float) (drone.getCoins() + coins);
+		float newPower = (float) (drone.getPower() + power);
 		double remainingCoins = 0;
 		double remainingPower = 0;
